requirements.txt
# Web Framework (Stable & Production-Ready)
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.3
pydantic-settings==2.2.1
python-multipart==0.0.9

# ML Framework - PyTorch with CUDA 12.4 (Stable LTS)
torch==2.2.1
torchvision==0.17.1
torchaudio==2.2.1

# Diffusers and Transformers (Latest Stable)
diffusers==0.27.2
transformers==4.38.2
accelerate==0.27.2
safetensors==0.4.2

# xformers for memory optimization (Compatible with PyTorch 2.2.1)
xformers==0.0.25

# Image Processing (Stable)
pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.4

# Background Removal (Stable)
rembg==2.0.56

# Utilities (Stable)
python-dotenv==1.0.1
aiofiles==23.2.1
huggingface-hub==0.21.4

# Logging (Stable)
loguru==0.7.2